Move result struct into its own file
